# openapi/bss-mvp-api-v1.yaml
openapi: 3.1.0
info:
  title: BSS MVP API
  version: 1.1.0
  summary: Implemented Backend MVP contract for the frozen BSS Frontend v1.0.0.
  description: |
    Implemented versioned contract for the functional Backend MVP; deployment remains a separate
    operational decision. The server derives organization,
    role and department scope from the verified session/device identity. Client-supplied
    scope is never authorization evidence. CSV and XLSX are the currently approved export
    formats. PDF is approved as a professional business export in v1.1.
  x-bss-status: MVP_IMPLEMENTED
  x-bss-contract-reviewed-at: 2026-07-17
  x-bss-frontend-release: frontend-v1.0.0
  x-bss-baseline-commit: 91323c7cdbbbbf7b965c4926c94a11af6d31bf62
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /api/v1
    description: Same-origin application API
...
security:
  - sessionCookie: []
paths:
...
  /attendance:
    get:
      tags: [Attendance]
      operationId: listAttendance
      summary: List Attendance
      x-bss-roles: [admin, manager]
      parameters:
        - $ref: '#/components/parameters/FromDate'
        - $ref: '#/components/parameters/ToDate'
        - $ref: '#/components/parameters/DepartmentId'
        - $ref: '#/components/parameters/WorkerIdQuery'
        - name: attendanceStatus
          in: query
          schema:
            $ref: '#/components/schemas/AttendanceStatus'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '200':
          description: Attendance rows and server-authoritative totals inside effective scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttendancePage'
  /workers/{workerId}/attendance:
    get:
      tags: [Attendance]
      operationId: getWorkerAttendance
      summary: Get Worker Attendance
      x-bss-roles: [admin, manager, worker]
      description: Worker may request only self; server ignores client identity claims and checks the session.
      parameters:
        - $ref: '#/components/parameters/WorkerId'
        - $ref: '#/components/parameters/FromDate'
        - $ref: '#/components/parameters/ToDate'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Attendance rows and authoritative totals for the permitted worker.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttendancePage'
        '403':
          $ref: '#/components/responses/Forbidden'
  /approved-leave-calendar:
    get:
      tags: [Leave]
      operationId: listApprovedLeaveCalendar
      summary: List Privacy-safe Approved Leave Calendar
      x-bss-roles: [admin, manager, worker, accountant]
      description: Returns only employee name and approved annual-leave dates inside the configured organization, department or team visibility boundary.
      parameters:
        - $ref: '#/components/parameters/FromDate'
        - $ref: '#/components/parameters/ToDate'
      responses:
        '200':
          description: Privacy-safe approved annual leave calendar without notes, illness or other absence reasons.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovedLeaveCalendar'
        '403':
          $ref: '#/components/responses/Forbidden'
  /leave-requests:
    get:
      tags: [Leave]
      operationId: listLeaveRequests
      summary: List Leave Requests
      x-bss-roles: [admin, manager, worker, accountant]
      parameters:
        - $ref: '#/components/parameters/FromDate'
        - $ref: '#/components/parameters/ToDate'
        - $ref: '#/components/parameters/DepartmentId'
        - name: leaveStatus
          in: query
          schema:
            $ref: '#/components/schemas/RequestStatus'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '200':
          description: Leave requests inside effective role scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeaveRequestPage'
    post:
      tags: [Leave]
      operationId: createLeaveRequest
      summary: Create Leave Request
      x-bss-roles: [worker]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LeaveRequestWrite'
      responses:
        '201':
          description: Pending leave request with server-calculated working days.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeaveRequest'
        '422':
          $ref: '#/components/responses/ValidationFailed'
...
  /leave-requests/{requestId}/approve:
    post:
      tags: [Leave]
      operationId: approveLeaveRequest
      summary: Approve Leave Request
      x-bss-roles: [admin, manager]
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        $ref: '#/components/requestBodies/Decision'
      responses:
        '200':
          description: Approved request and updated revision.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeaveRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /leave-requests/{requestId}/reject:
    post:
      tags: [Leave]
      operationId: rejectLeaveRequest
      summary: Reject Leave Request
      x-bss-roles: [admin, manager]
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        $ref: '#/components/requestBodies/RequiredDecision'
      responses:
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '200':
          description: Rejected request with mandatory decision note.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeaveRequest'
  /leave-requests/{requestId}/cancel:
    post:
      tags: [Leave]
      operationId: cancelOwnLeaveRequest
      summary: Cancel Own Leave Request
      x-bss-roles: [worker]
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/IfMatch'
      responses:
        '200':
          description: Own pending request cancelled.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeaveRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /correction-requests:
    get:
      tags: [Corrections]
      operationId: listCorrectionRequests
      summary: List Correction Requests
      x-bss-roles: [admin, manager, worker]
      parameters:
        - $ref: '#/components/parameters/FromDate'
        - $ref: '#/components/parameters/ToDate'
        - name: correctionStatus
          in: query
          schema:
            $ref: '#/components/schemas/RequestStatus'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '200':
          description: Correction requests inside effective role scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CorrectionRequestPage'
    post:
      tags: [Corrections]
      operationId: createCorrectionRequest
      summary: Create Correction Request
      x-bss-roles: [worker]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CorrectionRequestWrite'
      responses:
        '201':
          description: Pending correction preserving original values.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CorrectionRequest'
        '422':
          $ref: '#/components/responses/ValidationFailed'
  /correction-requests/{requestId}/approve:
    post:
      tags: [Corrections]
      operationId: approveCorrectionRequest
      summary: Approve Correction Request
      x-bss-roles: [admin, manager]
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        $ref: '#/components/requestBodies/Decision'
      responses:
        '200':
          description: Atomic correction, updated attendance day and audit reference.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CorrectionDecisionResult'
        '409':
          $ref: '#/components/responses/Conflict'
  /correction-requests/{requestId}/reject:
    post:
      tags: [Corrections]
      operationId: rejectCorrectionRequest
      summary: Reject Correction Request
      x-bss-roles: [admin, manager]
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        $ref: '#/components/requestBodies/RequiredDecision'
      responses:
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '200':
          description: Rejected correction; original attendance remains unchanged.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CorrectionRequest'
  /correction-requests/{requestId}/cancel:
    post:
      tags: [Corrections]
      operationId: cancelOwnCorrectionRequest
      summary: Cancel Own Correction Request
      x-bss-roles: [worker]
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/IfMatch'
      responses:
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '200':
          description: Own pending correction request cancelled.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CorrectionRequest'
  /report-exports:
    get:
      tags: [Reports]
      operationId: listReportExports
      summary: List Report Exports
      x-bss-roles: [admin, manager, accountant]
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '200':
          description: Export history inside effective role scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportExportPage'
    post:
      tags: [Reports]
      operationId: createReportExport
      summary: Create Report Export
      x-bss-roles: [admin, manager, accountant]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportExportWrite'
      responses:
        '202':
          description: Versioned export job accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportExport'
        '422':
          $ref: '#/components/responses/ValidationFailed'
...
  /report-exports/{exportId}:
    get:
      tags: [Reports]
      operationId: getReportExport
      summary: Get Report Export
      x-bss-roles: [admin, manager, accountant]
      parameters:
        - $ref: '#/components/parameters/ExportId'
      responses:
        '200':
          description: Export status and short-lived download URL when ready.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportExport'
        '404':
          $ref: '#/components/responses/NotFound'
  /report-exports/{exportId}/download:
    get:
      tags: [Reports]
      operationId: downloadReportExport
      summary: Download Report Export
      x-bss-roles: [admin, manager, accountant]
      parameters:
        - $ref: '#/components/parameters/ExportId'
      responses:
        '200':
          description: Ready, unexpired report artifact generated from the authoritative dataset.
          headers:
            Content-Disposition:
              schema:
                type: string
            X-Content-SHA256:
              schema:
                type: string
                pattern: '^[a-f0-9]{64}$'
          content:
            text/csv:
              schema:
                type: string
                format: binary
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
            application/pdf:
              schema:
                type: string
                format: binary
        '404':
          $ref: '#/components/responses/NotFound'
  /audit-events:
    get:
      tags: [Audit]
      operationId: listAuditEvents
      summary: List Audit Events
      x-bss-roles: [admin]
      parameters:
        - $ref: '#/components/parameters/FromDate'
        - $ref: '#/components/parameters/ToDate'
        - name: module
          in: query
          schema:
            type: string
        - name: entityId
          in: query
          schema:
            type: string
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '200':
          description: Append-only audit events; no mutation endpoint exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditEventPage'
  /terminals:
    get:
      tags: [Terminals]
      operationId: listTerminals
      summary: List Terminals
      x-bss-roles: [admin, manager]
      responses:
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '200':
          description: Terminal state and latest heartbeat; manager is read-only.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Terminal'
  /terminals/pair:
    post:
      tags: [Terminals]
      operationId: pairTerminal
      summary: Pair Terminal
      x-bss-roles: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TerminalPairRequest'
      responses:
        '201':
          description: Paired terminal. Device credential is returned once and never logged.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TerminalPairResult'
        '422':
          $ref: '#/components/responses/ValidationFailed'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: bss_session
      description: Secure, HttpOnly, SameSite session cookie.
    deviceSignature:
      type: apiKey
      in: header
      name: X-BSS-Signature
      description: |
        HMAC-SHA256 v1. The terminal signs UTF-8 bytes of
        METHOD + "\n" + PATH + "\n" + SHA256_HEX(BODY) + "\n" + X-BSS-Timestamp + "\n" + X-BSS-Nonce.
        The signature is lowercase hexadecimal. Server resolves X-BSS-Device-Id inside the tenant-owned
        encrypted credential store, permits at most 300 seconds clock skew and atomically consumes each nonce.
  parameters:
...
    FromDate:
      name: from
      in: query
      required: true
      schema:
        type: string
        format: date
    ToDate:
      name: to
      in: query
      required: true
      schema:
        type: string
        format: date
...
    DepartmentId:
      name: departmentId
      in: query
      schema:
        type: string
        format: uuid
...
    WorkerIdQuery:
      name: workerId
      in: query
      schema:
        type: string
        format: uuid
...